deps: upgrade tree-sitter 0.26 with the python and go grammars
Supersedes and closes #11, #12 and #13, which bumped the runtime and each
grammar separately. Any one alone produces E0308 mismatched types,
because the runtime and the grammars have to agree; Dependabot raises
them one crate at a time and none of the three could compile on its own.

  tree-sitter        0.23 -> 0.26
  tree-sitter-python 0.23 -> 0.25
  tree-sitter-go     0.23 -> 0.25

tree-sitter-typescript stays at 0.23.2, which is its latest release. All
three grammars depend on tree-sitter-language ^0.1 — the decoupling
layer — so they interoperate with the newer runtime regardless of their
own version numbers.

One API change: Node::child now takes u32 while Node::child_count still
returns usize. Four recursive tree walks iterate children this way, so
the loop bound is cast once per loop rather than casting at each call.

Behaviour is covered by the existing suite: all 57 radar-scanner tests
pass, and those exercise real parsing across all three grammars —
TypeScript operation-aware (S2) extraction plus Python and Go field-path
(S1) scanning — not just compilation.

Verified: fmt, clippy -D warnings, 470 workspace tests.

+++ radar-scanner/src/lib.rs
@@ -168,7 +168,8 @@ fn collect(
             }
         }
     }
-    for i in 0..node.child_count() {
+    // tree-sitter 0.26: Node::child takes u32 while child_count returns usize.
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect(&child, source, lang, out);
         }
@@ -320,7 +321,8 @@ fn collect_api_calls_s2(
             }
         }
     }
-    for i in 0..node.child_count() {
+    // tree-sitter 0.26: Node::child takes u32 while child_count returns usize.
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_api_calls_s2(&child, source, lang, out);
         }
@@ -627,7 +629,8 @@ fn collect_ops(
             }
         }
     }
-    for i in 0..node.child_count() {
+    // tree-sitter 0.26: Node::child takes u32 while child_count returns usize.
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_ops(&child, source, lang, derived, http_records);
         }
@@ -662,7 +665,8 @@ fn collect_derived_fields(
             }
         }
     }
-    for i in 0..node.child_count() {
+    // tree-sitter 0.26: Node::child takes u32 while child_count returns usize.
+    for i in 0..node.child_count() as u32 {
         if let Some(child) = node.child(i) {
             collect_derived_fields(&child, source, lang, derived, out);
         }
